+++ src/components/Table/THead/tests/index.js
@@ -1,35 +1,20 @@
 import React from 'react';
-import * as testHelpers from '../../../../utils/test-helpers';
 import THead from '../index';
+import { shallow } from 'enzyme';
 
 describe('components/Table/THead', () => {
-  const TestTableHead = (
-    <table>
-      <THead testSection="example-test-section">
-        <tr></tr>
-      </THead>
-    </table>
-  );
-
   it('should render as a `thead`', () => {
-    const component = testHelpers.renderIntoDocument(TestTableHead);
-    const tableHeadNode = testHelpers.getNodeFromComponent(component).children[0];
-
-    expect(tableHeadNode.tagName).toBe('THEAD');
+    const component = shallow(<THead></THead>);
+    expect(component.type()).toBe('thead');
   });
 
   it('should render children', () => {
-    const component = testHelpers.renderIntoDocument(TestTableHead);
-    const tableHeadNode = testHelpers.getNodeFromComponent(component).children[0];
-    const tableRowNode = tableHeadNode.children[0];
-
-    expect(tableRowNode.tagName).toBe('TR');
+    const component = shallow(<THead><th></th></THead>);
+    expect(component.containsMatchingElement(<th></th>)).toBe(true);
   });
 
   it('should render with test section', () => {
-    const component = testHelpers.renderIntoDocument(TestTableHead);
-    const tableHeadNode = testHelpers.getNodeFromComponent(component).children[0];
-
-    testHelpers.expectTestSectionToExist(tableHeadNode, 'example-test-section');
+    const component = shallow(<THead testSection="goose"></THead>);
+    expect(component.is('[data-test-section="goose"]')).toBe(true);
   });
 });